Unknown error goes to stderr. Tests added.
In die_unknown_error(), the error was sent to stdout. Now to stderr.

Moved the tests:
 - test_missing_credentials
 - test_expired_credentials
 from test_cli to test_error_handling

Added testcase in test_missing_credentials().
Added test_unknow_ClientError_no_stack_trace()
